requirements.txt
numpy
pillow
matplotlib
scikit-image
opencv-python
cryptography
pycryptodome
